requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-gis>=1.2
psycopg2-binary>=2.9.9
django-environ>=0.11.2
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
python-geohash>=0.8.5
geopy>=2.4.0
django-ninja>=0.22.0
django-cors-headers>=4.3.0
whitenoise>=6.6.0
redis>=5.0
PyJWT>=2.8.0
